+++ builder/vmware/common/step_register.go
@@ -1,11 +1,10 @@
-package iso
+package common
 
 import (
 	"context"
 	"fmt"
 	"time"
 
-	vmwcommon "github.com/hashicorp/packer/builder/vmware/common"
 	"github.com/hashicorp/packer/helper/multistep"
 	"github.com/hashicorp/packer/packer"
 )
@@ -13,14 +12,15 @@ import (
 type StepRegister struct {
 	registeredPath string
 	Format         string
+	KeepRegistered bool
 }
 
 func (s *StepRegister) Run(_ context.Context, state multistep.StateBag) multistep.StepAction {
-	driver := state.Get("driver").(vmwcommon.Driver)
+	driver := state.Get("driver").(Driver)
 	ui := state.Get("ui").(packer.Ui)
 	vmxPath := state.Get("vmx_path").(string)
 
-	if remoteDriver, ok := driver.(vmwcommon.RemoteDriver); ok {
+	if remoteDriver, ok := driver.(RemoteDriver); ok {
 		ui.Say("Registering remote VM...")
 		if err := remoteDriver.Register(vmxPath); err != nil {
 			err := fmt.Errorf("Error registering VM: %s", err)
@@ -40,18 +40,17 @@ func (s *StepRegister) Cleanup(state multistep.StateBag) {
 		return
 	}
 
-	driver := state.Get("driver").(vmwcommon.Driver)
+	driver := state.Get("driver").(Driver)
 	ui := state.Get("ui").(packer.Ui)
-	config := state.Get("config").(*Config)
 
 	_, cancelled := state.GetOk(multistep.StateCancelled)
 	_, halted := state.GetOk(multistep.StateHalted)
-	if (config.KeepRegistered) && (!cancelled && !halted) {
+	if (s.KeepRegistered) && (!cancelled && !halted) {
 		ui.Say("Keeping virtual machine registered with ESX host (keep_registered = true)")
 		return
 	}
 
-	if remoteDriver, ok := driver.(vmwcommon.RemoteDriver); ok {
+	if remoteDriver, ok := driver.(RemoteDriver); ok {
 		if s.Format == "" || config.SkipExport {
 			ui.Say("Unregistering virtual machine...")
 			if err := remoteDriver.Unregister(s.registeredPath); err != nil {
