+++ builder/vmware/common/step_register_test.go
@@ -1,4 +1,4 @@
-package iso
+package common
 
 import (
 	"context"
